requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
wavelink>=3.2.0
yt-dlp>=2023.12.30
aiohttp>=3.9.1
python-socketio>=5.10.0 